<feat> : <Сделать отдельный подсчет комментариев, bean для StatsClient>

+++ main-service/src/main/java/ru/practicum/model/dto/EventFullDto.java
@@ -43,5 +43,5 @@ public class EventFullDto {
 
     Integer views;
 
-    Integer comments;
+    Long comments;
 }
