+++ main-service/src/main/java/ru/practicum/model/dto/EventShortDto.java
@@ -29,5 +29,5 @@ public class EventShortDto {
 
     Integer views;
 
-    Integer comments;
+    Long comments;
 }
